Parse the target to extract host header
The host header was set to $http_host in nginx previously. When in
proxy mode, the header is now parsed from the target url, such that
multi-reverse-proxy setups work.

To enable this feature the ENV variable CERYX_HOST_FROM_TARGET has to be
set to true.

Reasoning: If ceryx is a front-facing proxy the requests will have a
host header designated for the target address. If ceryx is behind a
proxy itself, the host header will be the domain name of ceryx. In this
case ceryx cannot know the correct hostname for the target service.
In this scenario the hostname may be inferred from the target URL.

Alternatively this could have been already achieved via headers on a
route. Setting the host header on a route overrides the host header
from the request.